requirements.txt
Flask==3.1.2
tensorflow==2.20.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
Pillow==11.3.0

